development main slider component
