+++ src/js/modules/index.js
@@ -1,2 +1,3 @@
-export { Slider } from "./slider";
+export { Slider } from "./slider/slider";
 export { VideoPlayer } from "./playVideo";
+export { MainSlider } from "./slider/slider-main";
